Update CurrentSensorCombo.py
Final work. Made it so it can plot multiple things with an easy addition, NB currently limited to 3 sub plots maximum, can probably change this when I add in temperature.
Also, NB, limited by the names that the CSV saves under, if that changes then this code will need to be updated with the new names.